Conflicts:
	src/3rdparty/masm/yarr/YarrJIT.cpp
	src/qml/compiler/qv4instr_moth.cpp
	src/quick/handlers/qquicksinglepointhandler_p.h
	src/quick/handlers/qquicktaphandler.cpp
	src/quick/items/context2d/qquickcontext2d.cpp

Done-With: Ulf Hermann <ulf.hermann@qt.io>

+++ src/quick/scenegraph/util/qsgatlastexture.cpp
@@ -517,8 +517,8 @@ Texture::Texture(Atlas *atlas, const QRect &textureRect, const QImage &image)
     , m_nonatlas_texture(nullptr)
     , m_has_alpha(image.hasAlphaChannel())
 {
-    float w = atlas->size().width();
-    float h = atlas->size().height();
+    qreal w = atlas->size().width();
+    qreal h = atlas->size().height();
     QRect nopad = atlasSubRectWithoutPadding();
     m_texture_coords_rect = QRectF(nopad.x() / w,
                                    nopad.y() / h,
